Move widgets forms inside editor dir
