+++ lib/assets/javascripts/cartodb3/editor/editor-widget-view.js
@@ -1,6 +1,6 @@
 var cdb = require('cartodb-deep-insights.js');
 var template = require('./editor-widget.tpl');
-var WidgetsFormContentView = require('../widgets-form/widgets-form-content-view');
+var WidgetsFormContentView = require('./widgets-form/widgets-form-content-view');
 
 /**
  * View for an individual widget definition.
