+++ lib/assets/javascripts/cartodb3/editor/widgets-form/schemas/data/widgets-form-category-data-schema-model.js
