+++ lib/assets/javascripts/cartodb3/editor/widgets-form/schemas/data/widgets-form-formula-data-schema-model.js
