+++ lib/assets/javascripts/cartodb3/editor/widgets-form/schemas/data/widgets-form-histogram-data-schema-model.js
