+++ lib/assets/javascripts/cartodb3/editor/widgets-form/schemas/data/widgets-form-time-series-data-schema-model.js
