+++ lib/assets/javascripts/cartodb3/editor/widgets-form/schemas/style/widgets-form-category-style-schema-model.js
