+++ lib/assets/javascripts/cartodb3/editor/widgets-form/schemas/style/widgets-form-formula-style-schema-model.js
