+++ lib/assets/javascripts/cartodb3/editor/widgets-form/schemas/style/widgets-form-histogram-style-schema-model.js
