+++ lib/assets/javascripts/cartodb3/editor/widgets-form/schemas/style/widgets-form-time-series-style-schema-model.js
