+++ lib/assets/javascripts/cartodb3/editor/widgets-form/schemas/widgets-form-view.js
