+++ lib/assets/javascripts/cartodb3/editor/widgets-form/widgets-form-factory.js
