+++ lib/assets/test/spec/cartodb3/editor/widgets-form/widgets-content-view.spec.js
@@ -1,5 +1,5 @@
- var WidgetsFormContentView = require('../../../../javascripts/cartodb3/widgets-form/widgets-form-content-view');
- var WidgetDefinitionModel = require('../../../../javascripts/cartodb3/data/widget-definition-model');
+ var WidgetsFormContentView = require('../../../../../javascripts/cartodb3/editor/widgets-form/widgets-form-content-view');
+ var WidgetDefinitionModel = require('../../../../../javascripts/cartodb3/editor/data/widget-definition-model');
 
  describe('widgets-form/widgets-content-view', function () {
    beforeEach(function () {
