+++ lib/assets/test/spec/cartodb3/editor/widgets-form/widgets-form-factory.spec.js
@@ -1,5 +1,5 @@
-var WidgetsFormFactory = require('../../../../javascripts/cartodb3/widgets-form/widgets-form-factory');
-var WidgetDefinitionModel = require('../../../../javascripts/cartodb3/data/widget-definition-model');
+var WidgetsFormFactory = require('../../../../../javascripts/cartodb3/editor/widgets-form/widgets-form-factory');
+var WidgetDefinitionModel = require('../../../../../javascripts/cartodb3/editor/data/widget-definition-model');
 
 describe('widgets-form/widgets-form-factory', function () {
   describe('formula form', function () {
